Setup basic user tables
