+++ api/app.py
@@ -5,12 +5,14 @@ from database.api_db import ApiDb
 from database.models.bot_table import BotTable # noqa
 from database.models.deal_table import DealTable # noqa
 from database.models.order_table import ExchangeOrderTable # noqa
+from database.models.user_table import UserTable # noqa
 
 
 def main():
 
     api_db = ApiDb()
     api_db.init_db()
+    api_db.init_users()
     api_db.create_dummy_bot()
     result = api_db.select_bot("BTCUSDT")
     print("Created bot", result)
