+++ api/user/routes.py
@@ -1,8 +1,9 @@
 from fastapi import APIRouter, Depends
+from api.user.models.user import CreateUser
 from tools.handle_error import json_response, json_response_error
-from user.models.user import User
-from user.schemas import LoginRequest, UserResponse, UserSchema
-from auth import oauth2_scheme, Token, decode_access_token
+from user.services.controller import Controller
+from user.schemas import LoginRequest, UserResponse
+from api.user.services.auth import oauth2_scheme, Token, decode_access_token
 
 user_blueprint = APIRouter()
 
@@ -13,7 +14,7 @@ def get(token: str = Depends(oauth2_scheme)):
     Get all users
     """
     decode_access_token(token)
-    return User().get()
+    return Controller().get()
 
 
 @user_blueprint.get("/user/{email}", response_model=UserResponse, tags=["users"])
@@ -21,7 +22,7 @@ def get_one(email):
     """
     Get user by email
     """
-    return User().get_one(email)
+    return Controller().get_one(email)
 
 
 @user_blueprint.post("/user/login", tags=["users"], response_model=Token)
@@ -30,7 +31,7 @@ def login(data: LoginRequest):
     Get an access_token to keep the user in session
     """
     try:
-        access_token, user_data = User().login(data)
+        access_token, user_data = Controller().login(data)
         return json_response(
             {
                 "message": "Successfully logged in",
@@ -46,20 +47,20 @@ def login(data: LoginRequest):
 
 
 @user_blueprint.post("/user/register", tags=["users"])
-def add(data: UserSchema):
+def add(data: CreateUser):
     """
     Create/register a new user
     """
-    return User().add(data)
+    return Controller().add(data)
 
 
 @user_blueprint.put("/user", tags=["users"])
-def edit(user: UserSchema):
+def edit(user: CreateUser):
     """
     Modify details of a user that already exists.
     If the user does not exist, it will return a JSON error message
     """
-    return User().edit(user)
+    return Controller().edit(user)
 
 
 @user_blueprint.delete("/user/{email}", tags=["users"])
@@ -67,4 +68,4 @@ def delete(email: str):
     """
     Delete a user by email
     """
-    return User().delete(email)
+    return Controller().delete(email)
